Adding the Login functionality

+++ base/views.py
@@ -1,5 +1,10 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.db.models import Q
+from django.contrib.auth.models import User
+from django.contrib.auth.decorators import login_required
+from django.contrib import messages
+from django.contrib.auth import authenticate, login, logout
 from .models import Room, Topic
 from .forms import RoomForm
 
@@ -10,6 +15,39 @@ from .forms import RoomForm
 #     {'id':2,'name': 'Design with me'},
 #     {'id':3,'name': 'Frontend developers'},
 # ]
+
+def loginPage(request):
+    # Avoid user relogining in 
+    if request.user.is_authenticated:
+        return redirect('home')
+    
+    if request.method == "POST":
+        username = request.POST.get('username')
+        password = request.POST.get('password')
+        
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, "User dose not  exist")
+        
+        user = authenticate(request, username=username, password=password)
+        
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, "Username or password dose not exist")
+            
+    context = {}
+    return render(request, 'base/login_register.html', context)
+    
+        
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+            
+            
 
 def home(request):
     q= request.GET.get('q') if request.GET.get('q') != None else ''
@@ -31,6 +69,8 @@ def room(request, pk):
     context = {'room':room}
     return render(request, "base/room.html", context)
 
+
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == "POST":
@@ -41,25 +81,32 @@ def createRoom(request):
     context = {'form':form}
     return render(request, "base/room_form.html", context)
 
-
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+    
+    if request.user != room.host:
+        return HttpResponse("You are not allowed here!!")
 
     if request.method  == "POST":
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
             form.save()
             return redirect('home')
+        
+    context = {'form':form}
+    return render(request, 'base/room_form.html', context)
 
-
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+    
+    if request.user != room.host:
+        return HttpResponse("You are not allowed here!!")
+    
     if request.method == "POST":
         room.delete()
         return redirect('home')
     return render(request, 'base/delete.html', {'obj':room})
 
-
-    context = {'form':form}
-    return render(request, 'base/room_form.html', context)
